refactor: business logic 분리
