+++ src/pages/Result/Result.js
@@ -1,14 +1,13 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
+import Styled from './Result.style';
 import Button from '../../components/Button/Button';
 import LottoNumberList from '../../components/LottoNumberList/LottoNumberList';
 import LottoNumberItem from '../../components/LottoNumberItem/LottoNumberItem';
 import Modal from '../../components/Modal/Modal';
-import { RANKING, RANKING_TABLE, WINNING_TABLE } from '../../constants';
-import { getIntersectionCount, initObject } from '../../utils';
-import Styled from './Result.style';
 import WinningTable from '../../components/WinningTable/WinningTable';
 import PageTitle from '../../components/PageTitle/PageTitle';
+import { getProfitRate, getWinningResult } from '../../services/Result';
 
 class Result extends Component {
   constructor(props) {
@@ -20,53 +19,6 @@ class Result extends Component {
 
     this.handleOpenDetail = this.handleOpenDetail.bind(this);
     this.handleCloseDetail = this.handleCloseDetail.bind(this);
-    this.getProfitRate = this.getProfitRate.bind(this);
-  }
-
-  getRanking(lotto, { winningNumberList, bonusNumber }) {
-    let matchCount = getIntersectionCount(lotto, winningNumberList);
-
-    if (matchCount === 5) {
-      const matchBonusCount = getIntersectionCount(lotto, [bonusNumber]);
-      matchCount += matchBonusCount && 0.5;
-    }
-
-    return RANKING_TABLE[matchCount] || RANKING.NO_PRIZE;
-  }
-
-  getWinningResult(lottoList, { winningNumber, bonusNumber }) {
-    const winningNumberList = Object.values(winningNumber);
-
-    let winningResult = initObject(
-      Object.values(RANKING).filter((ranking) => ranking !== RANKING.NO_PRIZE),
-      0
-    );
-
-    Object.values(lottoList).forEach((lotto) => {
-      const ranking = this.getRanking(lotto, { winningNumberList, bonusNumber });
-
-      if (ranking !== RANKING.NO_PRIZE) {
-        winningResult = {
-          ...winningResult,
-          [ranking]: winningResult[ranking] + 1,
-        };
-      }
-    });
-
-    return winningResult;
-  }
-
-  getProfitRate(winningResult, moneyInput) {
-    const total = Object.entries(winningResult).reduce((accumulator, currentValue) => {
-      const [ranking, winningCount] = currentValue;
-      return accumulator + WINNING_TABLE[ranking].PRIZE * winningCount;
-    }, 0);
-
-    if (total < moneyInput) {
-      return Math.floor((total / moneyInput - 1) * 100);
-    }
-
-    return Math.floor(total / moneyInput) * 100;
   }
 
   handleOpenDetail() {
@@ -82,8 +34,8 @@ class Result extends Component {
   render() {
     const { isModalOpen } = this.state;
     const { lottoList, moneyInput, winningNumber, bonusNumber } = this.props?.location?.state;
-    const winningResult = this.getWinningResult(lottoList, { winningNumber, bonusNumber });
-    const profitRate = this.getProfitRate(winningResult, moneyInput);
+    const winningResult = getWinningResult(lottoList, { winningNumber, bonusNumber });
+    const profitRate = getProfitRate(winningResult, moneyInput);
 
     return (
       <>
